+++ arbitragelab/hedge_ratios/johansen.py
@@ -27,7 +27,7 @@ def get_johansen_hedge_ratio(price_data: pd.DataFrame, dependent_variable: str) 
 
     # Construct a Johansen portfolio
     port = JohansenPortfolio()
-    port.fit(price_data)
+    port.fit(price_data, dependent_variable)
 
     X = price_data.copy()
     X.drop(columns=dependent_variable, axis=1, inplace=True)
@@ -35,11 +35,7 @@ def get_johansen_hedge_ratio(price_data: pd.DataFrame, dependent_variable: str) 
     y = price_data[dependent_variable].copy()
 
     # Convert to a format expected by `construct_spread` function and normalize such that dependent has a hedge ratio 1.
-    hedge_ratios = port.cointegration_vectors.iloc[0].to_dict()
-    for ticker, h in hedge_ratios.items():
-        if ticker != dependent_variable:
-            hedge_ratios[ticker] = -h / hedge_ratios[dependent_variable]
-    hedge_ratios[dependent_variable] = 1.0
+    hedge_ratios = port.hedge_ratios.iloc[0].to_dict()
 
     residuals = construct_spread(price_data, hedge_ratios=hedge_ratios, dependent_variable=dependent_variable)