+++ tests/test_mean_reversion.py
@@ -13,8 +13,7 @@ import warnings
 import pandas as pd
 
 from arbitragelab.cointegration_approach import (JohansenPortfolio, EngleGrangerPortfolio,
-                                                 get_half_life_of_mean_reversion,
-                                                 linear_trading_strategy, bollinger_bands_trading_strategy)
+                                                 get_half_life_of_mean_reversion)
 
 
 class TestCointegration(unittest.TestCase):
@@ -111,43 +110,9 @@ class TestCointegration(unittest.TestCase):
 
         port = JohansenPortfolio()
         port.fit(price_data=self.data[self.data.columns[:5]])
+
+        # Construct spread
         mean_reverting_port = port.construct_mean_reverting_portfolio(self.data[self.data.columns[:5]])
         half_life = get_half_life_of_mean_reversion(mean_reverting_port)
 
         self.assertAlmostEqual(half_life, 31.23, delta=1e-2)
-
-    def test_linear_strategy(self):
-        """
-        Tests Simple Linear trading strategy function output.
-        """
-
-        port = JohansenPortfolio()
-        port.fit(price_data=self.data[self.data.columns[:5]])
-        mean_reverting_port = port.construct_mean_reverting_portfolio(self.data[self.data.columns[:5]])
-
-        # Window used as default half-life of mean reversion
-        linear_1 = linear_trading_strategy(mean_reverting_port)
-        linear_2 = linear_trading_strategy(mean_reverting_port, sma_window=10, std_window=20)
-
-        self.assertAlmostEqual(linear_1['target_quantity'].mean(), -0.01, delta=1e-2)
-        self.assertAlmostEqual(linear_1.iloc[100]['target_quantity'], -1.009, delta=1e-2)
-
-        self.assertAlmostEqual(linear_2['target_quantity'].mean(), -0.0084, delta=1e-2)
-        self.assertAlmostEqual(linear_2.iloc[100]['target_quantity'], -0.24, delta=1e-2)
-
-    def test_bollinger_bands_strategy(self):
-        """
-        Tests Bollinger Bands trading strategy function output.
-        """
-
-        port = JohansenPortfolio()
-        port.fit(price_data=self.data[self.data.columns[:5]])
-        mean_reverting_port = port.construct_mean_reverting_portfolio(self.data[self.data.columns[:5]])
-
-        bbands_strategy = bollinger_bands_trading_strategy(mean_reverting_port)
-        self.assertAlmostEqual(bbands_strategy['target_quantity'].mean(), 0.043, delta=1e-2)
-        self.assertAlmostEqual(bbands_strategy.iloc[100]['target_quantity'], 0.0, delta=1e-2)
-
-        # Test error raise
-        with self.assertRaises(ValueError):
-            bollinger_bands_trading_strategy(mean_reverting_port, entry_z_score=1, exit_z_score=3)