+++ docs/source/changelog.rst
@@ -12,6 +12,12 @@ Changelog
 ..
     For Help: https://releases.readthedocs.io/en/latest/index.html
 
+* :feature:`72` Added Bollinger Band Strategy to the Trading Strategies Module.
+* :feature:`72` Created a separate Minimum Profit Trading Strategy in the Trading Strategies Module.
+* :feature:`72` Created a separate Multivariate Cointegration Trading Strategy in the Trading Strategies Module.
+* :support:`72` Added Bollinger Band Strategy documentation.
+* :support:`72` Added Minimum Profit Trading Strategy documentation.
+* :support:`72` Added Multivariate Cointegration Trading Strategy documentation.
 * :support:`71` Added Raw Docs for debugging to documentation.
 * :support:`71` Added UML Diagram for debugging to documentation.
 * :feature:`70` Updated requirements versions to the newest stable numpy, pandas etc.